usb: chipidea: Add delay time for pinctrl switch over
On some hardware when we switch from one pin state to another it is
desirable to have a delay time to allow for peripheral chips to take
action based on the pin-state we have set.

This patch adds support for a delay time to be specified when doing that
switch. The switch and delay is a pattern then that is functionally
decomposed into ci_platform_set_pin_state().

If no delay time is specified the pin-switch happens as before with zero
delay. If a delay time is specified the pin-state is applied and then
the specified delay delay is applied.

Cc: Loic Poulain <loic.poulain@linaro.org>
Cc: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>