doc: usb: ci-hdrc-usb2: Add pinctrl delay variable
After applying a new pin state it is depending on the platform sometimes
appropriate to delay before proceeding. A delay can allow for a peripheral
component, such as a USB Hub to latch and take action based on the
pin-state as presented by the SoC.

This patch introduces a new optional property pin-switch-delay-us. The
delay period will trigger a delay as specified in the DT if and only if
switching pinctrl state, otherwise this variable has no effect.

Cc: Loic Poulain <loic.poulain@linaro.org>
Cc: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>