pil: modem: Add an additional check before assigning PIL memory
In a case where we run a Modem SSR on a secure device, the modem
ramdump process will trigger an MBA which scribbles the MSS region.
In order for MBA to scribble the MSS region, the MPSS has to own
the memory.
When we run a Modem SSR on a secure device with ramdump collection
disabled, MBA is triggered during init_image, at which point the
MSS region is owned by APSS. This causes a crash.
To prevent this crash, we check if an SSR was triggered, and if
ramdump collection is disabled, if this condition is met, we
actively assign the MSS region to MPSS for the MBA process.

Signed-off-by: Jin Liang <quic_jilia@quicinc.com>
Change-Id: Icebe2bd618b216c0b57065a7a1b1c51cb747dc8d