cnss: Fix the mem leak when do pci rescan
This issue is casued by the cnss_wlan_pci_probe() called
when do pci rescan with echo 1 > /sys/bus/pci/rescan. And
in current design it will alloc the cnss_seg_info once this
function cnss_wlan_pci_probe() called.

So adding the cnss_seg_info check to make it will not be
alloced if it is not NULL, and then this mem leak will not
happen when do pci rescan.

CRs-Fixed: 2726697
Change-Id: I2342abfe4a1eb8578bbd35df8a822d0c14ef5850
Signed-off-by: Ke Huang <keh@codeaurora.org>