net: ipa: use ipa_assert() for transactions
In gsi_trans_pool_alloc_common() there are two commented-out
assertions that communicate some assumptions about the count
value supplied.

We can't call ipa_assert() right there, because we need a device
pointer to do so.  One of its callers is gsi_trans_pool_alloc_dma(),
and it passes 1 as the count value, which meets the asserted
conditions.

The only other caller of gsi_trans_pool_alloc_common() is
gsi_trans_pool_alloc().  One of its callers also passes 1 as
count value, which is good.  In the other two callers, add
calls to ipa_assert() to verify the assumed conditions for
a validation build.

Signed-off-by: Alex Elder <elder@linaro.org>