net: ipa: implement MAX_READS_BEATS QSB data
Starting with IPA v4.0, a limit is placed on the number of bytes
outstanding in a transaction, to reduce latency.  The limit is
imposed only if this value is non-zero.

We currently don't set this, but in an upcoming patch we'll want to
set it impose a limit.  So allow a programmed value to be specified
in the platform configuration data.

Signed-off-by: Alex Elder <elder@linaro.org>