PCI: endpoint: Use blocking notifier instead of atomic
The use of atomic notifier causes sleeping in atomic context bug when
the EPC core functions are used in the notifier chain. This is due to the
use of EPC lock (mutex) in core functions protecting the concurrent use of
EPC.

So switch to blocking notifier to get rid of the bug as it runs in
non-atomic context and allows sleeping in notifier chain.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>