dmaengine: dw-edma: Use per-channel specfic direction in DMA slave caps
The eDMA driver doesn't have the platform support like devicetree. And the
channel allocation happens statically from index (0 to N-1) for write channels,
followed by the read channels continuously.

So for requesting a channel of a specific direction, the client drivers are
required to iterate over the list of available channels and find the apt one.
This can only be achieved when the channel specific direction is encoded in the
DMA slave capabilities.

But currently, the DMA slave capabilities contains the direction information
common to both read and write channels. Hence, by using "device_caps" callback,
the direction info of DMA slave cabalitites is overridden with the channel
specific one.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>