soc: qcom: spm: Add support for the CBF SPM on MSM8996
The msm8996 SoC contains several SPM instances, however most of them are
handled completely by the PSCI firmware. Add support for the CBF one,
the only one which needs to be programmed from the HLOS. For now it
contains just the init code. Later support for driving the SAW
regulators will be added.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>