arm64: dts: qcom: add SPM device nodes
Add SPM device nodes present on the msm8996 SoC.

The majority of them is added with the status equal to "reserved", as
these SPM devices are hanlded by the PSCI firmware. The devices are
added for documenting purpose and for the sake of earlier hardware
revisions which need to program them manually.

The only exception is the CBF SPM instance, which is going to be used
for APCC voltage scaling.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>