dt-bindings: msm: dsi-phy-28nm: Document missing #clock-cells dependency
We are missing #clock-cells in this binding. Add it in as a dependency with
a const value of 1 covering all current usages of this phy.

arch/arm/boot/dts/qcom-msm8974.dtsi
arch/arm/boot/dts/qcom-apq8064.dtsi
arch/arm64/boot/dts/qcom/msm8916.dtsi

The #clock-cells is required because because the mdss-ctrl has an
assigned-clock-parent back to the DSI PHY.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>