clk: qcom: apcs-msm8996: register sys_apcs_aux as a fixed rate clock
There is little sense in registering sys_apcs_aux as a child of gpll0.
The PLL is always-on. And listing the gpll0 as a property of the apcs
would delay probing till the GCC has been probed (while we would like
for the apcs to be probed as early as possible).

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>